src/requirements.txt
google-api-python-client==2.92.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==1.5.3
